ep4/requirements.txt
typer[all]
python-dotenv
twilio
sendgrid
boto3
auto-py-to-exe